+++ test/expectations/update_triggers_on_comments.rb
@@ -2,8 +2,8 @@
 
 class UpdateTriggersOnComments < ActiveRecord::Migration[7.0]
   def up
-    create_trigger "comments_before_update_tr", <<~SQL
-      CREATE OR REPLACE FUNCTION comments_before_update_tr() RETURNS TRIGGER
+    create_trigger "comments_before_update_of_title_tr", <<~SQL
+      CREATE OR REPLACE FUNCTION comments_before_update_of_title_tr() RETURNS TRIGGER
       AS $$
         BEGIN
           UPDATE posts SET comments_count = comments_count + 1 WHERE id = NEW.post_id;
@@ -11,26 +11,26 @@ class UpdateTriggersOnComments < ActiveRecord::Migration[7.0]
         END
       $$ LANGUAGE plpgsql;
 
-      CREATE TRIGGER comments_before_update_tr
+      CREATE TRIGGER comments_before_update_of_title_tr
       BEFORE UPDATE OF title ON "comments"
       FOR EACH ROW
-      EXECUTE FUNCTION comments_before_update_tr();
+      EXECUTE FUNCTION comments_before_update_of_title_tr();
     SQL
 
-    drop_trigger "comments_before_update_tr", <<~SQL
-      DROP TRIGGER IF EXISTS comments_before_update_tr ON "comments";
-      DROP FUNCTION IF EXISTS comments_before_update_tr;
+    drop_trigger "comments_before_update_of_title_tr", <<~SQL
+      DROP TRIGGER IF EXISTS comments_before_update_of_title_tr ON "comments";
+      DROP FUNCTION IF EXISTS comments_before_update_of_title_tr;
     SQL
   end
 
   def down
-    drop_trigger "comments_before_update_tr", <<~SQL
-      DROP TRIGGER IF EXISTS comments_before_update_tr ON "comments";
-      DROP FUNCTION IF EXISTS comments_before_update_tr;
+    drop_trigger "comments_before_update_of_title_tr", <<~SQL
+      DROP TRIGGER IF EXISTS comments_before_update_of_title_tr ON "comments";
+      DROP FUNCTION IF EXISTS comments_before_update_of_title_tr;
     SQL
 
-    create_trigger "comments_before_update_tr", <<~SQL
-      CREATE OR REPLACE FUNCTION comments_before_update_tr() RETURNS TRIGGER
+    create_trigger "comments_before_update_of_title_tr", <<~SQL
+      CREATE OR REPLACE FUNCTION comments_before_update_of_title_tr() RETURNS TRIGGER
       AS $$
         BEGIN
           UPDATE posts SET comments_count = comments_count + 1 WHERE id = NEW.post_id;
@@ -38,10 +38,10 @@ class UpdateTriggersOnComments < ActiveRecord::Migration[7.0]
         END
       $$ LANGUAGE plpgsql;
 
-      CREATE TRIGGER comments_before_update_tr
+      CREATE TRIGGER comments_before_update_of_title_tr
       BEFORE UPDATE OF title ON "comments"
       FOR EACH ROW
-      EXECUTE FUNCTION comments_before_update_tr();
+      EXECUTE FUNCTION comments_before_update_of_title_tr();
     SQL
   end
 end