+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased]
+
+- `where_clause` is now also used to compare triggers
+- Add trigger's columns in generated SQL and name
+
 ## [0.1.1]
 
 - Relax pg version dependency